hrthrfh

+++ src/HistoriqueTransactions/Benchmark.java
@@ -11,6 +11,7 @@ public class Benchmark {
     public static HistoriqueTreeSet donneesTreeSet(int taille) {
         System.gc();
         Runtime rt = Runtime.getRuntime();
+        //Mémoire occupée avant la génération des données
         long memoireAvant = rt.totalMemory() - rt.freeMemory();
         long debut = System.nanoTime();
 
@@ -20,12 +21,14 @@ public class Benchmark {
         System.out.println("Temps de génération des données : " + (fin-debut)*1e-6 + " ms.");
         System.out.println("Temps d'ajout d'une transaction : " + ((fin-debut)*1e-6)/taille + " ms.");
 
+        //Mémoire occupée après la génération des données
         long memoireApres = rt.totalMemory() - rt.freeMemory();
+        //Différence entre les deux qui donne l'occupation de mémoire du jeu de données créé
         long resultatOctets = memoireApres - memoireAvant;
 
         System.out.println("\n--- Occupation Mémoire (TreeSet) ---");
         System.out.println("Poids total : " + resultatOctets / 1024 + " Ko");
-        System.out.println("Poids moyen par transaction : " + (double)resultatOctets / taille + " octets");
+        System.out.println("Poids moyen par transaction : " + resultatOctets / taille + " octets");
 
         return historique;
     }
@@ -38,6 +41,7 @@ public class Benchmark {
         System.out.println("Temps de comptage de " + taille + " les transactions : " + (fin - debut) * 1e-6 + " ms.");
     }
 
+    //Calcul du temps du nombre de rechercheId passé en paramètre
     public static void tempsRechercheId(HistoriqueTreeSet h, int taille) {
         long debut = System.nanoTime();
         for (int i = 1; i <= taille; i++) {
@@ -52,6 +56,7 @@ public class Benchmark {
 
     }
 
+    //Calcul du temps utilisé pour parcourir tout le jeu de données
     public static void tempsParcoursChronologique(HistoriqueTreeSet h, int taille){
         long debut = System.nanoTime();
         LocalDate dateD = LocalDate.now();
@@ -61,6 +66,7 @@ public class Benchmark {
         System.out.println("Temps pour parcourir " + taille + " transactions : " + (fin-debut)*1e-6 + " ms.");
     }
 
+    //Calcul du temps utilisé pour supprimer le nombre de transactions passés en paramètre
     public static void tempsAnnulerTransaction(HistoriqueTreeSet h, int taille) {
         long debut = System.nanoTime();
         for (int i = 1; i <= taille; i++) {
